pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hbb-sanger"
version = "0.1.0"
description = "HBB Sanger sequencing analysis pipeline for beta-thalassaemia and sickle cell disease"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.83",
    "plotly>=5.20",
    "streamlit>=1.35",
    "pydantic>=2.7",
    "typer>=0.12",
]

[project.scripts]
hbb-analyze = "cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
